Renamed package from pl.learing to pl.learning
